chore(front-client): code review
