+++ apps/front-client/src/@types/chapter.ts
@@ -1,0 +1,5 @@
+import { Chapter } from '@prisma/client';
+
+export interface ChapitersI {
+  chapiters: Chapter[];
+}
