+++ package.json
@@ -45,6 +45,7 @@
     "clsx": "^1.2.1",
     "deep-equal": "^2.2.2",
     "jsonwebtoken": "^9.0.1",
+    "moment": "^2.29.4",
     "nanoevents": "^8.0.0",
     "next": "13.3.0",
     "next-auth": "^4.23.0",
